+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-08-18 21:11:48
-// Commit: fix(test): prevent double tap duplicate action in test
+// Updated at: 2026-08-18 22:17:36
+// Commit: fix(test): handle null assertion failure in test parser
 
 class testComponentState {
-  final String commitRef = '3722';
-  final String lastMessage = 'fix(test): prevent double tap duplicate action in test';
-  final DateTime timestamp = DateTime.parse('2026-08-18T21:11:48+05:00');
+  final String commitRef = '3726';
+  final String lastMessage = 'fix(test): handle null assertion failure in test parser';
+  final DateTime timestamp = DateTime.parse('2026-08-18T22:17:36+05:00');
 }